requirements.txt
sentence-transformers==4.1.0
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain-openai
langchain-google-genai
langchain-community
langchain-experimental
huggingface_hub==0.12.0
-e .